Split defi code from balances
[ui tests]
