+++ frontend/app/src/store/balances/actions.ts
@@ -1,28 +1,14 @@
 import { ActionTree } from 'vuex';
 import { currencies } from '@/data/currencies';
 import { FiatBalance } from '@/model/blockchain-balances';
-import {
-  BlockchainMetadata,
-  createTask,
-  ExchangeMeta,
-  taskCompletion,
-  TaskMeta
-} from '@/model/task';
+import { BlockchainMetadata, createTask, ExchangeMeta } from '@/model/task';
 import { TaskType } from '@/model/task-type';
 import {
-  convertMakerDAOVaults,
   convertManualBalances,
-  convertSupportedAssets,
-  convertVaultDetails
+  convertSupportedAssets
 } from '@/services/converters';
 import { api } from '@/services/rotkehlchen-api';
-import {
-  ApiMakerDAOVault,
-  ApiMakerDAOVaultDetails,
-  ApiManualBalance,
-  Watcher
-} from '@/services/types-api';
-import { WatcherTypes } from '@/services/types-common';
+import { ApiManualBalance } from '@/services/types-api';
 import { BalanceState } from '@/store/balances/state';
 import { notify } from '@/store/notifications/utils';
 import { Message, RotkehlchenState } from '@/store/store';
@@ -264,22 +250,7 @@ export const actions: ActionTree<BalanceState, RotkehlchenState> = {
     commit('ethAccounts', updateEth);
     commit('btcAccounts', updateBtc);
   },
-  async fetchDSRBalances({ commit }) {
-    const { task_id } = await api.dsrBalance();
-    const task = createTask(task_id, TaskType.DSR_BALANCE, {
-      description: `Fetching DSR Balances`,
-      ignoreResult: false
-    });
-    commit('tasks/add', task, { root: true });
-  },
-  async fetchDSRHistory({ commit }) {
-    const { task_id } = await api.dsrHistory();
-    const task = createTask(task_id, TaskType.DSR_HISTORY, {
-      description: `Fetching DSR History`,
-      ignoreResult: false
-    });
-    commit('tasks/add', task, { root: true });
-  },
+
   async fetchSupportedAssets({ commit, state }) {
     if (state.supportedAssets.length > 0) {
       return;
@@ -348,110 +319,6 @@ export const actions: ActionTree<BalanceState, RotkehlchenState> = {
         'setMessage',
         {
           title: 'Deleting Manual Balance',
-          description: `${e.message}`
-        } as Message,
-        { root: true }
-      );
-    }
-  },
-
-  async fetchWatchers({ commit, rootState: { session } }) {
-    if (!session?.premium) {
-      return;
-    }
-
-    try {
-      const watchers = await api.watchers();
-      commit('watchers', watchers);
-    } catch (e) {
-      notify(`Error: ${e}`, 'Fetching watchers', Severity.ERROR);
-    }
-  },
-
-  async addWatcher(
-    { commit },
-    watchers: Omit<Watcher<WatcherTypes>, 'identifier'>[]
-  ) {
-    const updatedWatchers = await api.addWatcher(watchers);
-    commit('watchers', updatedWatchers);
-    return updatedWatchers;
-  },
-
-  async deleteWatcher({ commit }, identifiers: string[]) {
-    const updatedWatchers = await api.deleteWatcher(identifiers);
-    commit('watchers', updatedWatchers);
-    return updatedWatchers;
-  },
-
-  async editWatcher({ commit }, watchers: Watcher<WatcherTypes>[]) {
-    const updatedWatchers = await api.editWatcher(watchers);
-    commit('watchers', updatedWatchers);
-    return updatedWatchers;
-  },
-
-  async fetchMakerDAOVaults({
-    commit,
-    rootGetters: { 'tasks/isTaskRunning': isTaskRunning }
-  }) {
-    if (isTaskRunning(TaskType.MAKEDAO_VAULTS)) {
-      return;
-    }
-
-    try {
-      const { task_id } = await api.makerDAOVaults();
-      const task = createTask(task_id, TaskType.MAKEDAO_VAULTS, {
-        description: `Fetching MakerDAO Vaults`,
-        ignoreResult: false
-      });
-
-      commit('tasks/add', task, { root: true });
-
-      const { result: makerDAOVaults } = await taskCompletion<
-        ApiMakerDAOVault[],
-        TaskMeta
-      >(TaskType.MAKEDAO_VAULTS);
-      commit('makerDAOVaults', convertMakerDAOVaults(makerDAOVaults));
-    } catch (e) {
-      commit(
-        'setMessage',
-        {
-          title: 'MakerDAO Vaults',
-          description: `${e.message}`
-        } as Message,
-        { root: true }
-      );
-    }
-  },
-
-  async fetchMakerDAOVaultDetails({
-    commit,
-    rootState: { session },
-    rootGetters: { 'tasks/isTaskRunning': isTaskRunning }
-  }) {
-    if (!session?.premium || isTaskRunning(TaskType.MAKERDAO_VAULT_DETAILS)) {
-      return;
-    }
-
-    try {
-      const { task_id } = await api.makerDAOVaultDetails();
-      const task = createTask(task_id, TaskType.MAKERDAO_VAULT_DETAILS, {
-        description: `Fetching MakerDAO Vault Details`,
-        ignoreResult: false
-      });
-
-      commit('tasks/add', task, { root: true });
-
-      const { result: makerDAOVaultDetails } = await taskCompletion<
-        ApiMakerDAOVaultDetails[],
-        TaskMeta
-      >(TaskType.MAKERDAO_VAULT_DETAILS);
-
-      commit('makerDAOVaultDetails', convertVaultDetails(makerDAOVaultDetails));
-    } catch (e) {
-      commit(
-        'setMessage',
-        {
-          title: 'MakerDAO Vault details',
           description: `${e.message}`
         } as Message,
         { root: true }
